Resolves #1060
